+++ tests/unit/test_dcatus3_validator.py
@@ -1,5 +1,7 @@
 from pathlib import Path
 
+import pytest
+
 from harvester.utils.general_utils import (
     build_dcatus3_validator,
     normalize_dataset_identifier,
@@ -19,19 +21,34 @@ DCATUS3_COMPLETE_EXAMPLE = (
 )
 
 DATASET_REF = "https://resources.data.gov/dcat-us/3.0.0/definitions/dataset"
+DATASET_VALIDATOR = build_dcatus3_validator(DCATUS3_DEFINITIONS, root_ref=DATASET_REF)
+
+
+@pytest.fixture
+def valid_dcatus3_dataset() -> dict:
+    return {
+        "@type": "Dataset",
+        "title": "Test Dataset",
+        "description": "A valid DCAT-US 3.0 dataset.",
+        "identifier": "https://example.gov/datasets/one",
+        "publisher": {"@type": "Organization", "name": "Test Agency"},
+        "contactPoint": {
+            "@type": "Kind",
+            "fn": "Test Contact",
+            "hasEmail": "mailto:test@example.gov",
+        },
+    }
 
 
 class TestBuildDcatus3Validator:
     def test_dataset_root_ref_validates_single_dataset(self, valid_dcatus3_dataset):
         """With the dataset root ref, a single dataset dict validates standalone."""
-        validator = build_dcatus3_validator(DCATUS3_DEFINITIONS, root_ref=DATASET_REF)
-        assert validator.is_valid(valid_dcatus3_dataset)
+        assert DATASET_VALIDATOR.is_valid(valid_dcatus3_dataset)
 
     def test_official_complete_example_dataset_passes_validation(self):
         """Upstream DCAT-US complete dataset example validates against our schema."""
-        validator = build_dcatus3_validator(DCATUS3_DEFINITIONS, root_ref=DATASET_REF)
         dataset = open_json(DCATUS3_COMPLETE_EXAMPLE)
-        assert validator.is_valid(dataset)
+        assert DATASET_VALIDATOR.is_valid(dataset)
 
     def test_official_complete_example_identifier_not_harvestable_without_atid(self):
         """Object identifier without @id is not harvestable."""
@@ -40,9 +57,8 @@ class TestBuildDcatus3Validator:
 
     def test_dataset_root_ref_flags_missing_required_field(self, valid_dcatus3_dataset):
         """A dataset missing the mandatory contactPoint produces errors."""
-        validator = build_dcatus3_validator(DCATUS3_DEFINITIONS, root_ref=DATASET_REF)
         del valid_dcatus3_dataset["contactPoint"]
-        errors = list(validator.iter_errors(valid_dcatus3_dataset))
+        errors = list(DATASET_VALIDATOR.iter_errors(valid_dcatus3_dataset))
         assert errors
         assert any("contactPoint" in e.message for e in errors)
